Fix the int to double test

+++ src/include/enesim_renderer_rectangle.h
@@ -25,12 +25,12 @@
  */
 
 EAPI Enesim_Renderer * enesim_renderer_rectangle_new(void);
-EAPI void enesim_renderer_rectangle_width_set(Enesim_Renderer *p, double w);
-EAPI void enesim_renderer_rectangle_width_get(Enesim_Renderer *p, double *w);
-EAPI void enesim_renderer_rectangle_height_set(Enesim_Renderer *p, double h);
-EAPI void enesim_renderer_rectangle_height_get(Enesim_Renderer *p, double *h);
-EAPI void enesim_renderer_rectangle_size_set(Enesim_Renderer *p, double w, double h);
-EAPI void enesim_renderer_rectangle_size_get(Enesim_Renderer *p, double *w, double *h);
+EAPI void enesim_renderer_rectangle_width_set(Enesim_Renderer *p, unsigned int w);
+EAPI void enesim_renderer_rectangle_width_get(Enesim_Renderer *p, unsigned int *w);
+EAPI void enesim_renderer_rectangle_height_set(Enesim_Renderer *p, unsigned int h);
+EAPI void enesim_renderer_rectangle_height_get(Enesim_Renderer *p, unsigned int *h);
+EAPI void enesim_renderer_rectangle_size_set(Enesim_Renderer *p, unsigned int w, unsigned int h);
+EAPI void enesim_renderer_rectangle_size_get(Enesim_Renderer *p, unsigned int *w, unsigned int *h);
 EAPI void enesim_renderer_rectangle_corner_radius_set(Enesim_Renderer *p, double radius);
 EAPI void enesim_renderer_rectangle_corners_set(Enesim_Renderer *p, Eina_Bool tl, Eina_Bool tr, Eina_Bool bl, Eina_Bool br);
 EAPI void enesim_renderer_rectangle_top_left_corner_set(Enesim_Renderer *r, Eina_Bool rounded);
